Importing Unity
check out the examples, a lot of code has just been imported from milu,
and is still not working in unity, like the listCreate functions and others.

also still needed to be imported is the mysql managers, etc ...

but the examples already work and I think this is one step closer to
what m-master had in his serenity where you can call each object's method.

I went a small step further and implemented an "editing" method that
allows the user to edit the object's variables.

This can be used as a "settings" editor, to tune variables, like
maximum number or rows in a table, sort order, mysql login, etc, etc etc